+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/aip/reference/AipResourceReference.kt
@@ -1,5 +1,6 @@
 package io.kanro.idea.plugin.protobuf.aip.reference
 
+import com.intellij.codeInsight.completion.CompletionUtilCore
 import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.openapi.util.TextRange
@@ -12,14 +13,15 @@ import io.kanro.idea.plugin.protobuf.aip.AipOptions
 import io.kanro.idea.plugin.protobuf.lang.completion.AddImportInsertHandler
 import io.kanro.idea.plugin.protobuf.lang.completion.ComposedInsertHandler
 import io.kanro.idea.plugin.protobuf.lang.completion.SmartInsertHandler
-import io.kanro.idea.plugin.protobuf.lang.file.FileResolver
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufFileOption
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufMessageDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufStringValue
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.ProtobufElement
 import io.kanro.idea.plugin.protobuf.lang.psi.stringRangeInParent
+import io.kanro.idea.plugin.protobuf.lang.psi.stringValue
 import io.kanro.idea.plugin.protobuf.lang.psi.stub.index.ResourceTypeIndex
 import io.kanro.idea.plugin.protobuf.lang.psi.value
+import io.kanro.idea.plugin.protobuf.lang.root.ProtobufRootResolver
 
 class AipResourceReference(element: ProtobufStringValue) : PsiReferenceBase<ProtobufStringValue>(element) {
     override fun resolve(): PsiElement? {
@@ -46,14 +48,15 @@ class AipResourceReference(element: ProtobufStringValue) : PsiReferenceBase<Prot
         result: MutableList<Any>,
         elements: MutableSet<ProtobufElement>
     ): Array<Any> {
-        if (!pattern.endsWith("IntellijIdeaRulezzz")) return arrayOf()
-        val searchName = pattern.substringBefore("IntellijIdeaRulezzz")
-        val scope = FileResolver.searchScope(element)
+        if (!pattern.endsWith(CompletionUtilCore.DUMMY_IDENTIFIER_TRIMMED)) return arrayOf()
+        val searchName = pattern.substringBefore(CompletionUtilCore.DUMMY_IDENTIFIER_TRIMMED)
+        val scope = ProtobufRootResolver.searchScope(element)
         val matcher = PlatformPatterns.string().contains(searchName)
         return StubIndex.getInstance().getAllKeys(ResourceTypeIndex.key, element.project).asSequence().filter {
             matcher.accepts(it)
         }.flatMap {
-            StubIndex.getElements(ResourceTypeIndex.key, it, element.project, scope, ProtobufElement::class.java).asSequence()
+            StubIndex.getElements(ResourceTypeIndex.key, it, element.project, scope, ProtobufElement::class.java)
+                .asSequence()
         }.mapNotNull {
             if (it in elements) return@mapNotNull null
             result += lookup(it, true) ?: return@mapNotNull null
@@ -84,7 +87,7 @@ class AipResourceReference(element: ProtobufStringValue) : PsiReferenceBase<Prot
             }
             is ProtobufFileOption -> {
                 val resourceName =
-                    element.value(AipOptions.resourceTypeField)?.stringValue?.value() ?: return null
+                    element.value(AipOptions.resourceTypeField)?.stringValue() ?: return null
                 LookupElementBuilder.create(
                     resourceName
                 ).withLookupString(resourceName.substringAfterLast('/'))