+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/settings/ProtobufSettings.kt
@@ -1,27 +1,39 @@
 package io.kanro.idea.plugin.protobuf.lang.settings
 
-import com.intellij.openapi.components.PersistentStateComponent
+import com.intellij.openapi.components.BaseState
+import com.intellij.openapi.components.SimplePersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
+import com.intellij.openapi.util.ModificationTracker
+import com.intellij.util.xmlb.annotations.Attribute
+import com.intellij.util.xmlb.annotations.Tag
+import com.intellij.util.xmlb.annotations.XCollection
 
 @State(
     name = "ProtobufSettings",
     storages = [Storage("protobuf.xml")]
 )
-class ProtobufSettings : PersistentStateComponent<ProtobufSettings.State> {
-    private var currentState = State()
-
-    override fun getState(): State {
-        return currentState
+class ProtobufSettings : SimplePersistentStateComponent<ProtobufSettings.State>(State()), ModificationTracker {
+    override fun getModificationCount(): Long {
+        return stateModificationCount
     }
 
-    override fun loadState(state: State) {
-        currentState = state
+    @Tag("entry")
+    class ImportRootEntry(path: String? = null, common: Boolean = true) : BaseState() {
+        @get:Attribute
+        var path by string()
+
+        @get:Attribute
+        var common by property(true)
+
+        init {
+            this.path = path
+            this.common = common
+        }
     }
 
-    data class ImportRootEntry(var path: String = "", var common: Boolean = true)
-
-    data class State(
-        var importRoots: List<ImportRootEntry> = listOf()
-    )
+    class State : BaseState() {
+        @get:XCollection(propertyElementName = "roots", style = XCollection.Style.v2)
+        var importRoots by list<ImportRootEntry>()
+    }
 }