+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/util/Extension.kt
@@ -1,5 +1,8 @@
 package io.kanro.idea.plugin.protobuf.lang.util
 
+import com.intellij.openapi.module.Module
+import com.intellij.openapi.module.ModuleUtilCore
+import com.intellij.psi.PsiElement
 import com.intellij.psi.util.QualifiedName
 
 fun String.toQualifiedName(): QualifiedName {
@@ -7,7 +10,19 @@ fun String.toQualifiedName(): QualifiedName {
 }
 
 fun QualifiedName.removeCommonPrefix(other: QualifiedName?): QualifiedName {
-    return removeCommonPrefix(other ?: return this, 0)
+    other ?: return this
+    var prefix = 0
+    while (true) {
+        if (prefix >= this.componentCount) break
+        if (prefix >= other.componentCount) break
+
+        if (this.components[prefix] == other.components[prefix]) {
+            prefix++
+        } else {
+            break
+        }
+    }
+    return this.removeHead(prefix)
 }
 
 fun QualifiedName.matchesSuffix(suffix: QualifiedName): Boolean {
@@ -22,15 +37,6 @@ fun QualifiedName.matchesSuffix(suffix: QualifiedName): Boolean {
     return true
 }
 
-private fun QualifiedName.removeCommonPrefix(other: QualifiedName, index: Int): QualifiedName {
-    if (index >= this.componentCount) return this.removeHead(index - 1)
-    if (index >= other.componentCount) return this.removeHead(index - 1)
-    if (this.components[index] == other.components[index]) {
-        return removeCommonPrefix(other, index + 1)
-    }
-    return this.removeHead(index)
-}
-
 fun <T> List<T>.contentEquals(other: List<T>): Boolean {
     if (this.size != other.size) return false
     forEachIndexed { index, t ->
@@ -38,3 +44,6 @@ fun <T> List<T>.contentEquals(other: List<T>): Boolean {
     }
     return true
 }
+
+val PsiElement.module: Module?
+    get() = ModuleUtilCore.findModuleForPsiElement(this)